+++ NEWS.md
@@ -1,6 +1,7 @@
 ## packer 0.0.6.9000
 
 - Use `rstudio` to open files, if available.
+- Added `scaffold_ambiorix` function
 
 ## packer 0.0.5